requirements.txt
Markdown==3.3.4
markdown-full-yaml-metadata==2.1.0
PyYAML==5.4.1
